Variable Argument Handling
Extension registration may now use the variable argument functionality to control the arguments of A_SpawnItemEx when the extension actor is spawned.

An extra boolean is in place and must be true for the extra arguments to be used.  This is supplied as a string literal.

All but the flag arguments are supported; x/y/z offset/velocity, angle, fail chance, and tid.

Args must be supplied in order to be used but defaults are preserved.